Support pulling cluster_name via cluster-name
Some non-reactive charms which provide the elasticsearch relation use
the cluster-name field to provide the cluster name, rather than
cluster_name as this interface defines.

As charms exist in the charmstore which provide and consume via both
patterns, it makes sense to update the requires side of this interface
to fall back to cluster-name if there does not appear to be a
cluster_name field defined.